(#40) Clean. Group reducers with state for simple features
