+++ Assets/Examples/Countdown/CountdownCommands.cs
@@ -5,4 +5,9 @@ namespace Examples.Countdown
     public class StopCountdownCommand : ICommand
     {
     }
+
+    public class StartCountdownCommand : ICommand
+    {
+        public float Seconds { get; set; }
+    }
 }
