--- Assets/Examples/Countdown/CountdownReducer.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 6cea05cdd75a4a2abbecda5b69d43ebf
-timeCreated: 1649215640
