+++ Assets/Examples/Countdown/FluxityExampleServiceInstaller.cs
@@ -1,6 +1,5 @@
 using AIR.Flume;
 using AIR.Fluxity;
-using Examples.Countdown;
 using UnityEngine;
 
 namespace Examples.Countdown
@@ -11,7 +10,7 @@ namespace Examples.Countdown
         protected override void InstallServices(FlumeServiceContainer container)
         {
             container
-                .RegisterFluxity(x => 
+                .RegisterFluxity(x =>
                     x.Feature(new CountdownState())
                         .Reducer<StartCountdownCommand>(CountdownReducer.StartCountDown)
                         .Reducer<StopCountdownCommand>(CountdownReducer.StopCountDown))
