--- Assets/Examples/Countdown/StartCountdownCommand.cs
@@ -1,9 +1,0 @@
-using AIR.Fluxity;
-
-namespace Examples.Countdown
-{
-    public class StartCountdownCommand : ICommand
-    {
-        public float Seconds { get; set; }
-    }
-}
