--- Assets/Examples/Countdown/StopCountdownCommand.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 1eaeb4fb328349de973159ff07fdf1b3
-timeCreated: 1649215569
