--- Assets/Examples/DataCommand/CounterReducer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 73882a3299c36edbc854170127d31cd7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
