+++ Assets/Examples/DataCommand/CounterState.cs
@@ -4,4 +4,12 @@ namespace Examples.DataCommand
     {
         public int CurrentCount;
     }
+
+    public static class CounterReducer
+    {
+        public static CounterState Change(CounterState state, ChangeCountCommand command)
+        {
+            return new CounterState { CurrentCount = state.CurrentCount + command.Delta };
+        }
+    }
 }
