--- Assets/Examples/ObjectData/FluxityExampleInitializer.cs
@@ -1,13 +1,0 @@
-using AIR.Fluxity;
-using UnityEngine;
-
-namespace Examples.ObjectData
-{
-    [DefaultExecutionOrder(1)]
-    public class FluxityExampleInitializer : FluxityInitializer
-    {
-        protected override void Initialize()
-        {
-        }
-    }
-}
