--- Assets/Examples/ObjectData/FluxityExampleInitializer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dd7a5d17337e15547b39b52cbefd234c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
