+++ Assets/Examples/ObjectData/FluxityExampleServiceInstaller.cs
@@ -10,8 +10,8 @@ namespace Examples.ObjectData
         protected override void InstallServices(FlumeServiceContainer container)
         {
             container
-                .RegisterFluxity(x => 
-                    x.Feature(ObjectDataState.CreateDefault())
+                .RegisterFluxity(x =>
+                    x.Feature(ObjectDataState.Create())
                         .Reducer<AddObjectDataCommand>(ObjectDataReducer.AddData))
                 ;
         }
