--- Assets/Examples/ObjectData/ObjectDataReducer.cs
@@ -1,14 +1,0 @@
-using System.Linq;
-
-namespace Examples.ObjectData
-{
-    public static class ObjectDataReducer
-    {
-        public static ObjectDataState AddData(ObjectDataState state, AddObjectDataCommand command)
-        {
-            var guids = state.Guids;
-            guids.AddRange(command.GuidsToAdd);
-            return new ObjectDataState { Guids = guids.Distinct().ToList() };
-        }
-    }
-}
