--- Assets/Examples/ObjectData/ObjectDataReducer.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 410ee399348f94743a3844becb89908e
-timeCreated: 1648787398
