--- Assets/Examples/Simple/CounterReducer.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 0bb4ce74a10f46a7b167bffd5d936124
-timeCreated: 1648787398
