--- Assets/Examples/Simple/CounterReducerAlt.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 57d7f9d529334d728c4e8942d49a3e1f
-timeCreated: 1648787414
