--- Assets/Examples/Simple/IncrementCountCommand.cs
@@ -1,8 +1,0 @@
-using AIR.Fluxity;
-
-namespace Examples.Simple
-{
-    public class IncrementCountCommand : ICommand
-    {
-    }
-}
