--- Assets/Examples/Simple/IncrementCountCommand.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 8faec8c0c7bb40298e96c811559b3249
-timeCreated: 1648787442
