+++ Assets/Examples/Simple/SimpleCountCommands.cs
@@ -5,4 +5,8 @@ namespace Examples.Simple
     public class DecrementCountCommand : ICommand
     {
     }
+
+    public class IncrementCountCommand : ICommand
+    {
+    }
 }
