+++ Assets/Examples/Spinner/SpinState.cs
@@ -5,4 +5,21 @@ namespace Examples.Spinner
         public bool DoSpin;
         public float DegreesPerSecond;
     }
+
+    public static class SpinnerReducers
+    {
+        public static SpinState StartSpin(SpinState state, StartSpinCommand command)
+        {
+            return new SpinState
+            {
+                DegreesPerSecond = command.DegreesPerSecond,
+                DoSpin = true,
+            };
+        }
+
+        public static SpinState StopSpin(SpinState state, StopSpinCommand command)
+        {
+            return new SpinState { DoSpin = false };
+        }
+    }
 }
