+++ Assets/Examples/Spinner/SpinnerCommands.cs
@@ -2,6 +2,10 @@ using AIR.Fluxity;
 
 namespace Examples.Spinner
 {
+    public class StopSpinCommand : ICommand
+    {
+    }
+
     public class StartSpinCommand : ICommand
     {
         public float DegreesPerSecond { get; set; }
