--- Assets/Examples/Spinner/SpinnerReducers.cs
@@ -1,18 +1,0 @@
-namespace Examples.Spinner
-{
-    public static class SpinnerReducers
-    {
-        public static SpinState StartSpin(SpinState state, StartSpinCommand command)
-        {
-            return new SpinState {
-                DegreesPerSecond = command.DegreesPerSecond,
-                DoSpin = true,
-            };
-        }
-
-        public static SpinState StopSpin(SpinState state, StopSpinCommand command)
-        {
-            return new SpinState { DoSpin = false };
-        }
-    }
-}
