--- Assets/Examples/Spinner/SpinnerReducers.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 297aa9bae40648548b6251a7dd5550a9
-timeCreated: 1649642383
