--- Assets/Examples/Spinner/StartSpinCommand.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 7fd2eb536639420ea4d2c937e3e6f941
-timeCreated: 1649642296
