--- Assets/Examples/Spinner/StopSpinCommand.cs
@@ -1,8 +1,0 @@
-using AIR.Fluxity;
-
-namespace Examples.Spinner
-{
-    public class StopSpinCommand : ICommand
-    {
-    }
-}
